LayoutTests/fast/workers/storage/interrupt-database-sync.html

<html>
<head>
<script src="../resources/worker-util.js"></script>
<script>
var worker;

function log(message)
{
    document.getElementById("console").innerHTML += message + "<br>";
}

function finishTest()
{
    log("PASS: database operations interrupted.");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function terminateWorker()
{
    worker.terminate();
    waitUntilWorkerThreadsExit(finishTest)
}

function runTest()
{
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    worker = new Worker('resources/interrupt-database-sync.js');
    worker.onmessage = function(event) {
        if (event.data == "terminate")
            terminateWorker();
        else
            log(event.data);
    };
}
</script>
</head>

<body onload="runTest()">
This test makes sure that all sync database operations are immediately interrupted when the worker needs to terminate.
<pre id="console">
</pre>
</body>
</html>
